+++ app/src/main/java/io/github/yawnoc/strokeinput/MainActivity.java
@@ -32,30 +32,32 @@ public class MainActivity
   implements View.OnClickListener
 {
   @Override
-  protected void onCreate(Bundle savedInstanceState) {
+  protected void onCreate(final Bundle savedInstanceState) {
     
     super.onCreate(savedInstanceState);
     setContentView(R.layout.activity_main);
     
-    Button inputSettingsButton = findViewById(R.id.input_settings_button);
+    final Button inputSettingsButton =
+      findViewById(R.id.input_settings_button);
     inputSettingsButton.setOnClickListener(this);
     
-    Button switchKeyboardButton = findViewById(R.id.switch_keyboard_button);
+    final Button switchKeyboardButton =
+      findViewById(R.id.switch_keyboard_button);
     switchKeyboardButton.setOnClickListener(this);
     
   }
   
   @Override
-  public void onClick(View view) {
+  public void onClick(final View view) {
     
-    int viewId = view.getId();
+    final int viewId = view.getId();
     
     if (viewId == R.id.input_settings_button) {
       Intent intent = new Intent(Settings.ACTION_INPUT_METHOD_SETTINGS);
       startActivity(intent);
     }
     else if (viewId == R.id.switch_keyboard_button) {
-      InputMethodManager inputMethodManager =
+      final InputMethodManager inputMethodManager =
         (InputMethodManager) getSystemService(INPUT_METHOD_SERVICE);
       inputMethodManager.showInputMethodPicker();
     }