+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Fixed NullPointerException for `onSinglePointerTouchEvent` of null key
 - Added fillColour attribute for the keyboard as a whole
+- Lightened key fill and text colours slightly
 
 
 ## [v0.3.0] Basic keyboard behaviour (2021-07-08)