+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [Unreleased]
 
+
+## [v0.4.0] Symbols keyboard and fixes (2021-07-09)
+
 - Added symbols keyboard and implemented switching to it
 - Fixed NullPointerException for `onSinglePointerTouchEvent` of null key
 - Fixed key press colour change dependent on activity background
@@ -54,7 +57,9 @@ the actual functionality has not been implemented yet.
 
 
 [Unreleased]:
-  https://github.com/stroke-input/stroke-input-android/compare/v0.3.0...HEAD
+  https://github.com/stroke-input/stroke-input-android/compare/v0.4.0...HEAD
+[v0.4.0]:
+  https://github.com/stroke-input/stroke-input-android/compare/v0.3.0...v0.4.0
 [v0.3.0]:
   https://github.com/stroke-input/stroke-input-android/compare/v0.2.0...v0.3.0
 [v0.2.0]: